+++ src/Service/Models/StepUpVerifyRequest.cs
@@ -1,0 +1,3 @@
+namespace Passwordless.Service.Models;
+
+public record StepUpVerifyRequest(string Token, string Context);
